ConfigProvider/Config should be Descriptor/Describable
This brings the instantiation of Config in line with the rest of UI, and allows the Config implementations to enjoy the core taglib support and other existing infrastructure, such as form validation and checks. For example, Config object can now (and should) have default config.jelly that the rest of the client plugins can utilize.

Unfortunately, some of the ConfigProvider methods have already deviated from the Descriptor class, such that I couldn't eliminate some of the code duplication, but I still think this alignment is better long-term