Improved target path parsing for Config File.
Checking for "." anywhere in full filepath is a poor way of determining whether a path is to a file or a directory.  Naming conventions allow for periods in directory names and do not necessitate them in file names.

Plugin should assume target is specifying filename unless (1) path includes a trailing slash or (2) the path already exists in the workspace as a directory.