[JENKINS-46925] Support custom folder types.
It looks like the code was mostly working with AbstractFolder however
the important Action was looking for types of Folder.  This widens the
type to AbstractFolder.

This is not binary compatable with old code calling FolderConfigiFileAction,
but I could not find any occurences in the wild.

Any existing configuration should be re-read without issues.