[JENKINS-67122] Fix NPE in Pipeline Snippet Generator
The Pipeline snippet generator reports an exception when trying to
generate a sample using configFileProvider.  It assumes the context is
always not-null, yet in the context of the Pipeline syntax generator
at the root, there is no context to be used.

https://issues.jenkins.io/browse/JENKINS-67122

See the issue locally by opening
http://localhost:8080/jenkins/pipeline-syntax then selecting
configFileProvider from the drop down list

A different issue is visible when opening from inside a job definition.
It reports HTTP error 403.