Fix test failure by using a longer user name
The base64 credential masking was detecting the "bot" user name earlier
in the log and replacing it.  A longer user name that is not a substring
of other text in the log file prevents that.

Also extends the bot password text to a longer string to match the longer
user name string.