ui enhancement
having a separate table for each config file type has the problem that
the the column width are different for each type. Also the comment is
written in a separate line
This change shows everything in a single table with a tbody for each type.
The types are separated by some space and each type has a header with
title and column description. Each file is on one line.
Styling is moved to a dedicated css file.

remove inline style