fix readonly mode
setting var readOnlyMode in the show-config jelly files allows to
include edit-config directly. For those that add credentials check the
readOnlyMode variable and disable the delete and add buttons accordingly
Allow extended read for the dofill and docheck methods so a job config
can be viewed with extended read permissions.